+++ src/try_from_iterator.rs
@@ -1,5 +1,11 @@
+#[cfg(doc)]
+use std::collections::{BTreeMap, HashMap};
+
+#[cfg(doc)]
+use crate::KeyCollision;
+
 /// Allows converting an iterator into a container that may fail to be constructed.
-pub trait TryFromIterator<T> {
+pub trait TryFromIterator<T>: Sized {
     /// The error that may occur when converting the iterator into the container.
     type Error;
 
@@ -7,8 +13,8 @@ pub trait TryFromIterator<T> {
     ///
     /// # Example
     ///
-    /// Provided implementations for `HashMap`, `BTreeMap`, `hashbrown::HashMap`, and
-    /// `indexmap::IndexMap` all work similarly, and provide the first instance of a
+    /// Provided implementations for [`HashMap`], [`BTreeMap`], [`hashbrown::HashMap`], and
+    /// [`indexmap::IndexMap`] all work similarly, and provide the first instance of a
     /// colliding key within the [`KeyCollision`] error.
     ///
     /// ```rust
@@ -21,23 +27,5 @@ pub trait TryFromIterator<T> {
     /// let result = HashMap::try_from_iter([(1, 2), (2, 3)]);
     /// assert_eq!(result, Ok(HashMap::from([(1, 2), (2, 3)])));
     /// ```
-    fn try_from_iter<I: IntoIterator<Item = T>>(into_iter: I) -> Result<Self, Self::Error>
-    where
-        Self: Sized;
-}
-
-/// An error indicating that a duplicate key was found in the provided data.
-#[derive(Debug, thiserror::Error, derive_more::Constructor, PartialEq, Eq, PartialOrd, Ord)]
-#[error("Key collision")]
-pub struct KeyCollision<K> {
-    /// The key that caused the collision.
-    pub key: K,
-}
-
-/// An error indicating that a duplicate value was found in the provided data.
-#[derive(Debug, thiserror::Error, derive_more::Constructor, PartialEq, Eq, PartialOrd, Ord)]
-#[error("Value collision")]
-pub struct ValueCollision<T> {
-    /// The value that caused the collision.
-    pub value: T,
+    fn try_from_iter<I: IntoIterator<Item = T>>(into_iter: I) -> Result<Self, Self::Error>;
 }